Split challenge rated tool to a separate repo
