--- app/challenge-rated/_lib/CreatureToggleOptions.ts
@@ -1,7 +1,0 @@
-import type { RadioOption } from '@/app/lib/types';
-
-const CreatureToggleOptions: RadioOption[] = [
-  { displayText: 'Enemies', value: 0 },
-  { displayText: 'Allies', value: 1 },
-];
-export default CreatureToggleOptions;
