+++ app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from 'next';
 import React from 'react';
 
-import Footer from './components/Footer';
-import Header from './components/Header';
+import Footer from './ui/Footer';
+import Header from './ui/Header';
+
 import './styles/dracula.css';
 import './styles/global.css';
