--- app/lib/types.ts
@@ -1,4 +1,0 @@
-export type RadioOption = {
-  displayText: string;
-  value: number;
-};
