+++ app/ui/Header/styles.module.css
@@ -11,7 +11,6 @@
   background: #282a36;
   height: 50px;
   padding: 5px;
-  position: fixed;
   display: flex;
   align-items: center;
   box-shadow: rgba(0, 0, 0, 0.15) 0px 5px 10px 0px;
