+++ app/ui/icons/IconTrash.tsx
